Seperate importer to parital classes. Add IsAnimationNode property in SceneNode
