+++ Source/HelixToolkit.SharpDX.Assimp.Shared/HelixToolkit.SharpDX.Assimp.Shared.projitems
@@ -11,5 +11,8 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Extensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Importer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImporterPartial_Animation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImporterPartial_Mesh.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ImporterPartial_Material.cs" />
   </ItemGroup>
 </Project>
